requirements.txt
# Core ML/DL frameworks
torch>=2.0.0
torchvision>=0.15.0

# Data processing
numpy>=1.23.0,<2.0.0
pandas>=1.5.0,<3.0.0

# Visualization
matplotlib>=3.6.0,<4.0.0
seaborn>=0.12.0,<1.0.0

# Image processing
Pillow>=9.3.0,<11.0.0

# Machine learning utilities
scikit-learn>=1.2.0,<2.0.0

# Progress bars
tqdm>=4.64.0,<5.0.0
